fixed refresh and state problems by adding observers to UserSingleton
