+++ src/components/JobList.tsx
@@ -1,12 +1,12 @@
 import { useState } from "react";
 import { JobListing } from "../data/job_listing";
 import { ListingView } from "./ListingView";
-import { UserSingleton, waitForUser } from "../data/user";
+import { UserSingleton } from "../data/UserSingleton";
 
 export function JobList({ listings }: { listings: JobListing[] }): JSX.Element {
   const [refresh, setRefresh] = useState(false);
-
-  waitForUser(UserSingleton.getInstance(), refresh, setRefresh, "JobList", false);
+  
+  UserSingleton.getInstance().addListener(refresh, setRefresh, "JobList");
 
   return (
     <div className="feedBox">
